requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
django-autocomplete-light==3.9.7
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==19.6.1
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==10.0.0
PyDecimal==1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
whitenoise==6.5.0
